### 2020-04-22 #31 @bruAristimunha
[Re] Deep Convolution Neural Network and Autoencoders-Based Unsupervised Feature Learning of EEG Signals
**Original article:**
T. Wen and Z. Zhang, "[Deep Convolution Neural Network and Autoencoders-Based Unsupervised Feature Learning of EEG Signals](https://ieeexplore.ieee.org/abstract/document/8355473)," in IEEE Access, vol. 6, pp. 25399-25410, 2018.

**Scientific domain:** Neuroscience, EEG, Computational Neuroscience
@rougier , @ogrisel , @qihongl ,
@rth , @rgutzen , @emmanuelle , @jpgard
