### 2020-06-02 #31 @bruAristimunha
Hi @rougier,
Sorry to bother you, is it possible to switch to another possible editor? It seems to me that he cannot take over.
