### 2020-06-17 #31 @neuronalX
Hi @gdetor
I am not an expert in CNNs and given the number of experiments, the paper would need some expert to review the codes properly.

May be you could ask to @vthierry or Thalita Firmo Drumond.
