### 2020-08-14 #31 @koustuvsinha
Here is my review. The reproducibility effort is very comprehensive and thorough, but the report is difficult to read and needs more work and polish to be in a shape to present in the journal. I'm happy to reply and re-evaluate the updated draft.

- In Section 3.1, the disagreement in the literature on the total number of instances is not clear. I think this relates to the Bonn University EEG dataset, which is missing from this section.
- I think authors can drop the Related works section, and cut down on the discussion of the autoencoder model. These details can be referred to in the original paper.
- I propose to rewrite Section 5.1 - this is about choosing the channels in the dataset to work on, with the hypothesis that EEG channels with higher variance would offer more discriminatory signal for epilepsy prediction. It wasn't immediately clear on why authors report "three interpretations". On deeper inspection, it seems variance calculation can be done in multiple ways: 1. for each sample. 2. per person or 3. for the full dataset. This needs to be mentioned explicitly.
    - I think the 1st one (for each sample) is the process used by the original authors
    - The original authors did report choosing the channels based on evidence from "first 10 patients". Did you also notice that?
- In Section 5.2, What is "dimension"? 64 vs 128? Not explained.
- Since there is a lot of analysis results presented in the paper, I would suggest the authors to clearly denote the analysis in their own section. For each section, start with which claim of the original paper you investigated, and follow with your results. Dumping all tables doesn't make the paper readable, and we don't understand which points of the original paper was or not reproducible.
- Table 15 to 21 - too many values, could this be aggregated?
- What are the key takeways? Summary of reproduction needed.
- Conclusion: Typo of "Auto-Enconder"?

I have not gone through the code in detail yet. I would wait for the changes, atleast a better logical presentation of the results before verifying the claims in the code.
