### 2020-08-15 #31 @benureau
@gdetor: sorry for the delay, going through the review now, assembling the hardware I need to run the code.

@bruAristimunha: did you contact the original authors about what seemed ambiguous in the original paper? Specifically, for instance, section 5.1? And all the assumptions you made in section 3.1? While a paper should be replicable without reaching out to the original authors, there's a point to be made about ferreting out the missing details and making them explicit in a replication effort submitted to ReScience.
