### 2020-09-01 #31 @bruAristimunha
I apologize for the delay in responding, @gdetor, @koustuvsinha and @benureau. All the reviewers' comments were accepted. The work will go through re-editing with the removal of several sections that are not related to reproduction itself.

We will try to improve the fluidity of the text, explaining that we did not have a response from the authors and that we did not obtain the same result, in terms of variance and results.

This review round will be conducted by @DiogoEduardo. In particular, we plan to remove sections: 2 (Related Work), 3.2 (AutoEncoders), 4.3 (metrics definition) and 5.3 (Extension). In addition to improved fluidity as a whole.
