### 2020-09-02 #31 @benureau
@bruAristimunha: I am halfway through my review. I will wait for you to update your submission, since the changes you are preparing seem important.

A few comments:
- Have you noticed that in the original paper, Table 3 and 4 contain the exact same values? With the loss and the dataset changing between the two? It seems to be a copy-paste error. It would affect the validity of some of the comparisons you are making.
- You are using different terminology for the classifiers than the original article. Be sure to explicitly link them together "AE-CDNN-MAE is [11]'s AE-CDNN-L1", etc.
- Being explicit in which Table of the original article is being reproduced or compared in each of the figure's captions would help avoid the reader second-guessing anything.
- The lack of variance of the SVMs in Tables 2 and 3 of the reproduction is concerning, would suggest a qualitative difference between your reproduction and the original article's implementation. It should probably be discussed, not just stated.

Code:
- jupyter is not in the requirements.txt file.
- Give the command to correctly launch the jupyter notebook, and from which directory level.
- Installing currently throws an error, because the scipy version is not specified:
```
ERROR: After October 2020 you may experience errors when installing or updating packages. This is because pip will change the way that it resolves dependency conflicts.

We recommend you use --use-feature=2020-resolver to test your packages with the new resolver before it becomes the default.

tensorflow 2.1.0 requires scipy==1.4.1; python_version >= "3", but you'll have scipy 1.5.2 which is incompatible.
```
- The third cell of 7.4.5 in the notebook complains about pyarrow dependency missing.
- The last cell of 3. throws errors with True on all three first lines:
```bash
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: ` rm -r $"{}/data/chbmit/variance_accumulated".format(base_fold)'
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: ` rm -r $"{}/data/chbmit/variance_file".format(base_fold)'
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: ` rm -r $"{}/data/chbmit/variance_person".format(base_fold)'
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: ` rm -r $"{}/data/chbmit/reduced".format(base_fold)'
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: ` rm -r $"{}/data/boon/reduced".format(base_fold)'
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: ` rm -r $"{}/data/chbmit/save_model".format(base_fold)'
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: ` rm -r $"{}/data/boon/save_model".format(base_fold)'
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: ` rm -r $"{}/data/boon".format(base_fold)'
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: ` rm -r $"{}/data/chbmit".format(base_fold)'
mkdir: cannot create directory ‘../article/figure’: File exists
mkdir: cannot create directory ‘../article/table’: File exists
```
- I was able to re-generate all the figures present in your article from your pre-saved training
- I got a 404 error when invoking `download_chbmit` to download from scratch.
```
Downloading the folder information: ../data/chbmit/
Downloading the folder files: ../data/chbmit/
Downloading the folder information: ../data/chbmit/chb01/
Downloading the folder files: ../data/chbmit/chb01/
Downloading the folder information: ../data/chbmit/chb02/
Downloading the folder files: ../data/chbmit/chb02/

HTTPError                                 Traceback (most recent call last)
[...]
```
- Currently re-running training. Will post my results when it finishes.
