### 2020-09-09 #31 @benureau
After running into some trouble with the `rm -r` commands above not working and having to do it manually, I was unable to reproduce Figure 2:
```python
<ipython-input-25-8de2b88f16b3> in <module>
      1 variance_per_person = get_variance_by_person(PATH_CHBMIT,
----> 2                                              range_= (1,11))

~/rescience/ReScience-submission/code/variance.py in get_variance_by_person(path_dataset, range_)
    318             var_pearson.append(accumulate_var)
    319
--> 320         variance_df = DataFrame(var_pearson).drop("time", 1)
    321
    322         variance = DataFrame(

~/.pyenv/versions/miniconda3-latest/envs/eeg/lib/python3.7/site-packages/pandas/core/frame.py in __init__(self, data, index, columns, dtype, copy)
    472                     if is_named_tuple(data[0]) and columns is None:
    473                         columns = data[0]._fields
--> 474                     arrays, columns = to_arrays(data, columns, dtype=dtype)
    475                     columns = ensure_index(columns)
    476

~/.pyenv/versions/miniconda3-latest/envs/eeg/lib/python3.7/site-packages/pandas/core/internals/construction.py in to_arrays(data, columns, coerce_float, dtype)
    466     elif isinstance(data[0], ABCSeries):
    467         return _list_of_series_to_arrays(
--> 468             data, columns, coerce_float=coerce_float, dtype=dtype
    469         )
    470     elif isinstance(data[0], Categorical):

~/.pyenv/versions/miniconda3-latest/envs/eeg/lib/python3.7/site-packages/pandas/core/internals/construction.py in _list_of_series_to_arrays(data, columns, coerce_float, dtype)
    514         index = getattr(s, "index", None)
    515         if index is None:
--> 516             index = ibase.default_index(len(s))
    517
    518         if id(index) in indexer_cache:

TypeError: object of type 'int' has no len()
```
and figure 1 and 2 were different than the one in the paper, with for instance much less channels on Figure 1. I stopped there. Could you make sure that all the figures are reproducible from scratch, on a vanilla machine/vanilla conda environment? Or explain the sources of possible differences?

Obtained figure 1:
![figure1](https://user-images.githubusercontent.com/2088743/92559260-2f975e00-f2ab-11ea-8e09-875264a0ee02.png)

Obtained figure 3:
![figure3](https://user-images.githubusercontent.com/2088743/92559309-4c339600-f2ab-11ea-85cb-ce4f9af0c310.png)
