### 2020-09-19 13344f9
reviewers changes
Reviewers comments solved:

In Section 3.1, the disagreement in the literature on the total number of instances is not clear. I think this relates to the Bonn University EEG dataset, which is missing from this section.
I think authors can drop the Related works section, and cut down on the discussion of the autoencoder model. These details can be referred to in the original paper.
I propose to rewrite Section 5.1 - this is about choosing the channels in the dataset to work on, with the hypothesis that EEG channels with higher variance would offer more discriminatory signal
[…]
