+++ article/table/accuracy_chbmit_maae-reproduction.tex
@@ -12,6 +12,6 @@
 \textbf{128} &       0.6500 &      0.8100 &      0.8275 &         0.7600 &         0.8100 &       0.8075 &     0.7775 &       0.8300 &  0.784062 \\ \hline
 \end{tabular}
 }
-\caption{Accuracy values obtained by the same methodology - CHBMIT Dataset with MAAE.}
+\caption{Classification results of MAAE for dataset 2 - Table 5 in [11].}
 \label{table:accuracy_chbmit_maae-reproduction}
 \end{table}
