+++ article/table/acc-mae-boon-fold-2.tex
@@ -10,6 +10,6 @@
 5      &          0.4 &         0.6 &         0.6 &            0.6 &            0.6 &          0.6 &        0.6 &          0.4 \\ \hline
 \end{tabular}
 }
-\caption{Accuracy in Classification, with loss MAE, on each fold cross-validation, for Dataset 1.}
+\caption{Classification results of MAE for dataset 1 - Table 6 in [11].}
 \label{table:acc-mae-boon-fold-2}
 \end{table}
