+++ article/table/accuracy_boon_maae-reproduction.tex
@@ -12,6 +12,6 @@
 \textbf{128} &        0.790 &       0.616 &        0.72 &          0.726 &          0.758 &        0.750 &      0.772 &        0.620 &  0.71900 \\ \hline
 \end{tabular}
 }
-\caption{Accuracy values obtained by the same methodology - Boon Dataset with MAAE.}
+\caption{Classification results of MAAE for dataset 1 - Table 3 in [11].}
 \label{table:accuracy_boon_maae-reproduction}
 \end{table}
