+++ article/table/acc-mae-chbmit-fold-2.tex
@@ -10,6 +10,6 @@
 5      &       0.6875 &      0.5000 &      0.5000 &         0.7000 &         0.6375 &       0.6250 &     0.7000 &        0.700 \\ \hline
 \end{tabular}
 }
-\caption{Accuracy in Classification, with loss MAE, on each fold cross-validation, for Dataset 2.}
+\caption{Classification results of MAE for dataset 2 - Table 8 in [11].}
 \label{table:acc-mae-chbmit-fold-2}
 \end{table}
