+++ article/table/accuracy_chbmit_mae-reproduction.tex
@@ -12,6 +12,6 @@
 \textbf{128} &       0.6475 &      0.8200 &      0.8425 &         0.7275 &         0.8000 &       0.8275 &     0.7950 &       0.7975 &  0.782187 \\ \hline
 \end{tabular}
 }
-\caption{Accuracy values obtained by the same methodology - CHBMIT Dataset with MAE.}
+\caption{Classification results of MAE for dataset 2 - Table 4 in [11].}
 \label{table:accuracy_chbmit_mae-reproduction}
 \end{table}
