+++ article/article.tex
@@ -1,6 +1,6 @@
 \documentclass{rescience}
 \usepackage[utf8,latin1]{inputenc}
-\usepackage{lipsum}
+	\usepackage{lipsum}
 \usepackage{graphicx}
 \begin{document}
 \input{header.tex}
