+++ article/table/acc-maae-boon-fold-2.tex
@@ -10,6 +10,6 @@
 5      &         0.71 &         0.6 &         0.6 &           0.68 &           0.73 &          0.6 &       0.74 &         0.54 \\ \hline
 \end{tabular}
 }
-\caption{Accuracy in Classification, with loss MAAE, on each fold cross-validation, for Dataset 1.}
+\caption{Classification results of MAAE for dataset 1 - Table 7 in [11].}
 \label{table:acc-maae-boon-fold-2}
 \end{table}
