+++ article/table/metrics_boon_cv_10.tex
@@ -7,6 +7,6 @@
 32        &        0.802 &         0.6 &       0.602 &          0.722 &          0.828 &        0.712 &      0.792 &        0.638 &   0.7120 \\
 \end{tabular}
 }
-\caption{Accuracy in Classification, in the first dataset with $CV=10$.}
+\caption{Accuracy in Classification, in dataset 1 with $CV=10$ - Table 9 in [11].}
 \label{table:metrics_boon_cv_10}
 \end{table}
