+++ article/table/accuracy_boon_mae-reproduction.tex
@@ -12,6 +12,6 @@
 \textbf{128} &        0.780 &       0.716 &       0.754 &          0.700 &          0.786 &        0.792 &      0.782 &        0.616 &  0.74075 \\ \hline
 \end{tabular}
 }
-\caption{Accuracy values obtained by the same methodology - Boon Dataset with MAE.}
+\caption{Classification results of MAE for dataset 1 - Table 2 in [11].}
 \label{table:accuracy_boon_mae-reproduction}
 \end{table}
