+++ requirements.txt
@@ -5,6 +5,6 @@ numpy==1.18.2
 pandas==1.0.3
 scikit_learn==0.22.2
 seaborn==0.10.0
-tensorflow==2.1.0
+tensorflow==2.1.2
 tqdm==4.38.0
 wget==3.2
