### 2020-10-05 1e8ccaf
Cuts and minor corrections
Removed all the extra metrics sections (only accuracy maintained), corrected minor errors in the text.
