+++ article/content.tex
@@ -1,23 +1,13 @@
 
-This paper presents our efforts to reproduce the results achieved by the authors of the original article. We follow the steps and models described in their article and the same public data sets of EEG Signals. Epilepsy affects more than 65 million people globally, and EEG Signals are critical to analyze and recognize epilepsy. Although the efforts in the last years, it is still challenging to extract useful information from these signals and select useful features in a diagnostic application. We construct a deep convolution network and autoencoders-based model (AE-CDNN) in order to perform unsupervised feature learning. We use the AE-CDNN to extract the features of the available data sets, and then we use some common classifiers to classify the features. The results obtained demonstrate that the proposed AE-CDNN outperforms the traditional feature extraction based classification techniques by achieving better accuracy of classification.
+This paper presents our efforts to reproduce the results achieved by the authors of the original article \cite{WenZha:2018}. We follow the steps and models described in their article and the same public data sets of EEG Signals. Epilepsy affects more than 65 million people globally, and EEG Signals are critical to analyze and recognize epilepsy. Although the efforts in the last years, it is still challenging to extract useful information from these signals and select useful features in a diagnostic application. We construct a deep convolution network and autoencoders-based model (AE-CDNN) in order to perform unsupervised feature learning. We use the AE-CDNN to extract the features of the available data sets, and then we use some common classifiers to classify the features. The results obtained demonstrate that the proposed AE-CDNN outperforms the traditional feature extraction based classification techniques by achieving better accuracy of classification.
 
 \section{Introduction}
 
-Epilepsy is a chronic neurological disorder, and it is becoming one of the most common neurological diseases in the world \cite{most_commum:2002}. Approximately $1\%$ of the world's population is affected by epilepsy representing more than 65 million people affected \cite{global-epilepsy:2019,Epilepsia:2010}. This disorder is characterized by the occurrence of spontaneous convulsions due to the abnormal synchronous firing of the cortical neurons \cite{stafstrom2015seizures}. This physical reaction can generate many problems for patients, including physical harm caused by the loss of consciousness, shame and discrimination \cite{thomas2011confronting}.
+The paper {\it Deep Convolution Neural Network and Autoencoders-Based Unsupervised Feature Learning of EEG Signals} \cite{WenZha:2018} presents an unsupervised approach to learn from EEG signals of epilepsy patients. It uses dimension reduction algorithms to extract features since the original data is high-dimensional. Then, it uses different common classifiers to classify the features obtained. 	
 
-Frequent seizures are dangerous conditions because, at the moment of disruption of the body can occur falls, fractures, burns, car accidents, and other serious physical injuries \cite{mollaouglu2013injuries}. Epilepsy can be defined as a permanent predisposition in the brain to cause epileptic seizures \cite{stafstrom2015seizures}.
+We re-implement the author's method from scratch as best as we could by using only the paper as instruction. In addition, we get new results by combining the proposal classifiers into a classifier by set voting, and we add new metrics.
 
-To be diagnosed with epilepsy, the patient must have at least two seizures that are caused by comorbidities known in the medical literature, such as: extremely low blood sugar \cite{schauwecker2012effects}. Even when correctly diagnosed and treated, the epileptic patient still suffers side effects and sporadic seizures. The epileptic seizures can cause even irreversible damage to the brain, and then we can visualize the importance of analyzing epilepsy to improve the life quality and the medical treatments for these patients \cite{stafstrom2015seizures}. 
-
[…]
