### 2020-10-05 deaa810
Section 3.3 removed
