+++ article/content.tex
@@ -5,7 +5,7 @@ This paper presents our efforts to reproduce the results achieved by the authors
 
 The paper {\it Deep Convolution Neural Network and Autoencoders-Based Unsupervised Feature Learning of EEG Signals} \cite{WenZha:2018} presents an unsupervised approach to learn from EEG signals of epilepsy patients. It uses dimension reduction algorithms to extract features since the original data is high-dimensional. Then, it uses different common classifiers to classify the features obtained. 	
 
-We re-implement the author's method from scratch as best as we could by using only the paper as instruction. In addition, we get new results by combining the proposal classifiers into a classifier by set voting, and we add new metrics.
+We re-implement the author's method from scratch as best as we could by using only the paper as instruction. In addition, we get new results by combining the proposal classifiers into a classifier by set voting.
 
 This paper is organized as follows: Section \ref{sec:propose} introduces the methodological proposal employed, and their differences from \cite{WenZha:2018}. Section \ref{sec:metho} lists the experimental validation process using epilepsy datasets. Section \ref{sec:resu} presents the corresponding results and analyzes our approach. Finally, conclusions were summarized in Section \ref{sec:concl}.
 
