+++ article/article.tex
@@ -1,7 +1,10 @@
 \documentclass{rescience}
-\usepackage[utf8,latin1]{inputenc}
- 	\usepackage{lipsum}
+\usepackage[utf8]{inputenc}
+\usepackage{lipsum}
 \usepackage{graphicx}
+\usepackage{caption}
+\usepackage{subcaption}
+
 \begin{document}
 \input{header.tex}
 \input{content.tex}
