+++ article/content.tex
@@ -11,52 +11,47 @@ This paper is organized as follows: Section \ref{sec:propose} introduces the met
 
 
 
-
-
-
-
 \section{Methodology Proposal}
 \label{sec:propose}
 
 In this section, we describe implementation details, as the core is the reproducible aspect of our reference article. We introduce the idea and implementation of autoencoder/feature learning and our version of the model in \cite{WenZha:2018}, explaining the differences we have made in the original model.
 
-In our study, we keep the autoencoder and feature learning as proposed. However, in the classification, in addition to the individual classifiers, we also employ a large ensemble learning classifier, which decides by majority vote the object class.
-
 \subsection{Implementation Details}\label{subsec:implementation}
 
-We decided to reproduce the implementation described in the article using Keras \cite{chollet2018keras} and backend in TensorFlow \cite{tensorflow}. Our repository includes a list of all the required libraries employed in acquiring the datasets and running the model (the original and the proposed one). According to the methodology proposed in \cite{Fuente:2019}, we store all the checkpoints for the trained models, for reproduction purposes. Besides that, the training logs can be visualized using TensorBoard tool.
+We decided to reproduce the implementation described in the article using Keras \cite{chollet2018keras} with TensorFlow  \cite{tensorflow} as backend. Our repository includes a list of all the required libraries employed in acquiring the datasets and running the model (the original and the proposed one). Using to the methodology proposed in \cite{Fuente:2019}, we store all the checkpoints for the trained models for reproduction purposes. Besides that, the training logs can be visualized using the ensorBoard tool.
 
-Given the lack of information about implementation in the original paper, some assumptions or cuts are made: 
+Given the lack of information about some implementation details in the original paper, some assumptions or cuts are made: 
 
 \begin{itemize}
 \item The number of epoch in the AutoEnconder is assumed to be $5000$; 
 \item The number of samples per batch size is assumed to be $256$; 
-\item A column of the Bonn University EEG dataset was removed since the authors of \cite{WenZha:2018} used $4096$ instances and in this database there is $4097$ instances. The removed attribute is at the endpoint of each object; 
+\item The last column of the Bonn University EEG dataset was removed since the authors of \cite{WenZha:2018} used $4096$ features and in this database there are $4097$ column. 
 \item In the Children's Hospital of Boston EEG database we use the channel reported by the author to train the AutoEncoder;
-\item The loss function presented in equation $12$ in \cite{WenZha:2018} was implemented and we also compared the result obtained with \textit{MAPE}; 
-\item The value of the seeds selected in all classifiers, data splitting and elsewhere was $42$; 
-\item The train-validation ratio was $80\%-20\%$ to AutoEnconder, in classifiers we use cross-validation with 5 or 10-fold; 
-\item Given a sizing problem, we resized the values using the MinMax method, before the classification process.
+\item The loss function presented in equation $12$ in \cite{WenZha:2018} was implemented; 
+\item The value of the seeds used in all classifiers, data splitting and elsewhere was $42$; 
+\item The train-validation ratio was $80\%-20\%$ for the to AutoEnconder. For the classifiers we used 5 or 10-fold cross-validation 
+\item We resized the values using the MinMax method, before the classification process.
 \item The classifier presented in the final subsection (NN2) was not reproduced for lack of information; 
 \end{itemize}
 
-The experiments were performed using a CPU with Intel Core i7-5930K with 3.50 GHz and two GPUs: Nvidia Quadro K5200 and GeForce GTX 970. Some experiments were also run using Nvidia Titan X.
[…]
