+++ requirements.txt
@@ -5,6 +5,8 @@ numpy==1.18.2
 pandas==1.0.3
 scikit_learn==0.22.2
 seaborn==0.10.0
-tensorflow==2.1.2
+tensorflow-gpu==2.3.0
 tqdm==4.38.0
 wget==3.2
+scipy==1.5.0
+pyarrow==2.0.0
