### 2021-06-04 #31 @bruAristimunha
Hello @gdetor and @rougier

I apologize for the delay in contacting you. Unfortunately, for personal reasons, I won't be able to handle the modifications. However, the second author @DiogoEduardo will continue with the necessary revisions.
