+++ article/table/specificity_boon_maae-reproduction.tex
@@ -1,18 +1,20 @@
-\begin{table}[!ht]
+\begin{table}
 \centering
-\resizebox{\textwidth}{!}{
-\begin{tabular}{c|cccccccccc}
-\textbf{Dimension} & \textbf{k\_neighbors} & \textbf{svm\_linear} & \textbf{svm\_radial} & \textbf{decision\_tree} & \textbf{random\_forest} & \textbf{multi\_layer} &\textbf{ada\_boost} &\textbf{gaussian\_nb} &\textbf{ensemble} & \textbf{average} \\ \hline
-\textbf{2        } &        0.690 &       0.000 &       0.000 &          0.605 &          0.655 &        0.000 &      0.690 &        0.940 &     0.575 &  0.461667 \\
-\textbf{4        } &        0.690 &       0.000 &       0.000 &          0.635 &          0.700 &        0.000 &      0.665 &        0.900 &     0.610 &  0.466667 \\
-\textbf{8        } &        0.650 &       0.000 &       0.000 &          0.605 &          0.610 &        0.000 &      0.695 &        0.765 &     0.540 &  0.429444 \\
-\textbf{16       } &        0.830 &       0.000 &       0.000 &          0.840 &          0.795 &        0.065 &      0.775 &        0.985 &     0.800 &  0.565556 \\
-\textbf{32       } &        0.845 &       0.000 &       0.000 &          0.785 &          0.795 &        0.040 &      0.800 &        0.980 &     0.775 &  0.557778 \\
-\textbf{64       } &        0.870 &       0.000 &       0.000 &          0.835 &          0.835 &        0.585 &      0.750 &        0.980 &     0.875 &  0.636667 \\
-\textbf{128      } &        0.845 &       0.190 &       0.905 &          0.805 &          0.755 &        0.765 &      0.730 &        0.945 &     0.915 &  0.761667 \\
-\textbf{256      } &        0.780 &       0.565 &       0.905 &          0.750 &          0.740 &        0.760 &      0.755 &        0.940 &     0.910 &  0.789444 \\ \hline
-\end{tabular}
-}
 \caption{Specificity values obtained by the same methodology - boon Dataset with maae.}
-\label{table:specificity_boon_maae-reproduction}
+\label{specificity_boon_maae-reproduction}
+\begin{tabular}{lrrrrrrrrrr}
+\toprule
+\textbf{name\_classifier} &  k\_neighbors &  svm\_linear &  svm\_radial &  decision\_tree &  random\_forest &  multi\_layer &  ada\_boost &  gaussian\_nb &  ensemble &   average \\
+\textbf{Dimension} &              &             &             &                &                &              &            &              &           &           \\
+\midrule
+\textbf{2        } &        0.675 &        0.00 &       0.000 &          0.650 &          0.700 &        0.000 &      0.795 &        0.780 &     0.600 &  0.466667 \\
+\textbf{4        } &        0.715 &        0.00 &       0.000 &          0.675 &          0.705 &        0.000 &      0.665 &        0.945 &     0.615 &  0.480000 \\
+\textbf{8        } &        0.865 &        0.00 &       0.000 &          0.770 &          0.810 &        0.000 &      0.780 &        0.975 &     0.770 &  0.552222 \\
+\textbf{16       } &        0.815 &        0.00 &       0.000 &          0.760 &          0.805 &        0.010 &      0.725 &        0.980 &     0.745 &  0.537778 \\
+\textbf{32       } &        0.850 &        0.00 &       0.510 &          0.705 &          0.790 &        0.680 &      0.775 &        0.970 &     0.885 &  0.685000 \\
[…]
