+++ article/table/f-measure_chbmit_maae-reproduction.tex
@@ -1,18 +1,20 @@
-\begin{table}[!ht]
+\begin{table}
 \centering
-\resizebox{\textwidth}{!}{
-\begin{tabular}{c|cccccccccc}
-\textbf{Dimension} & \textbf{k\_neighbors} & \textbf{svm\_linear} & \textbf{svm\_radial} & \textbf{decision\_tree} & \textbf{random\_forest} & \textbf{multi\_layer} &\textbf{ada\_boost} &\textbf{gaussian\_nb} &\textbf{ensemble} & \textbf{average} \\ \hline
-\textbf{2} &     0.698258 &    0.009756 &    0.251173 &       0.663391 &       0.684510 &     0.463725 &   0.648123 &     0.705439 &  0.631161 &  0.528393 \\
-\textbf{4} &     0.811006 &    0.320984 &    0.680609 &       0.805932 &       0.811680 &     0.664390 &   0.799736 &     0.767600 &  0.753578 &  0.712835 \\
-\textbf{8} &     0.826095 &    0.321100 &    0.584106 &       0.749253 &       0.797223 &     0.637806 &   0.796049 &     0.722898 &  0.699157 &  0.681521 \\
-\textbf{16} &     0.818304 &    0.661183 &    0.803174 &       0.805980 &       0.844897 &     0.788908 &   0.814428 &     0.852531 &  0.819065 &  0.800941 \\
-\textbf{32} &     0.724628 &    0.697587 &    0.790896 &       0.746017 &       0.811508 &     0.789309 &   0.807596 &     0.836740 &  0.782786 &  0.776341 \\
-\textbf{64} &     0.715778 &    0.698471 &    0.819348 &       0.774569 &       0.824374 &     0.796561 &   0.816922 &     0.855324 &  0.815164 &  0.790723 \\
-\textbf{128} &     0.463376 &    0.793962 &    0.815910 &       0.747477 &       0.804942 &     0.801733 &   0.765678 &     0.797050 &  0.808231 &  0.755373 \\
-\textbf{256} &     0.492048 &    0.808966 &    0.829264 &       0.734030 &       0.784070 &     0.845676 &   0.802964 &     0.812416 &  0.824258 &  0.770410 \\ \hline
+\caption{F-measure values obtained by the same methodology - chbmit Dataset with maae.}
+\label{f-measure_chbmit_maae-reproduction}
+\begin{tabular}{lrrrrrrrrrr}
+\toprule
+\textbf{name\_classifier} &  k\_neighbors &  svm\_linear &  svm\_radial &  decision\_tree &  random\_forest &  multi\_layer &  ada\_boost &  gaussian\_nb &  ensemble &   average \\
+\textbf{Dimension} &              &             &             &                &                &              &            &              &           &           \\
+\midrule
+\textbf{2        } &     0.665833 &    0.052174 &    0.474252 &       0.727899 &       0.649296 &     0.616910 &   0.693557 &     0.679058 &  0.668949 &  0.580881 \\
+\textbf{4        } &     0.706794 &    0.009756 &    0.220965 &       0.706509 &       0.661048 &     0.458205 &   0.694987 &     0.686022 &  0.640786 &  0.531675 \\
+\textbf{8        } &     0.736381 &    0.212734 &    0.594080 &       0.738242 &       0.762492 &     0.665565 &   0.733423 &     0.669964 &  0.684347 &  0.644137 \\
+\textbf{16       } &     0.755031 &    0.681081 &    0.823232 &       0.744578 &       0.821072 &     0.798301 &   0.790609 &     0.839119 &  0.810461 &  0.784832 \\
[…]

[…]
