+++ article/table/precision_chbmit_mae-reproduction.tex
@@ -1,20 +1,18 @@
-\begin{table}
+\begin{table}[!ht]
 \centering
-\caption{Precision values obtained by the same methodology - chbmit Dataset with mae.}
-\label{precision_chbmit_mae-reproduction}
-\begin{tabular}{lrrrrrrrrrr}
-\toprule
-\textbf{name\_classifier} &  k\_neighbors &  svm\_linear &  svm\_radial &  decision\_tree &  random\_forest &  multi\_layer &  ada\_boost &  gaussian\_nb &  ensemble &   average \\
-\textbf{Dimension} &              &             &             &                &                &              &            &              &           &           \\
-\midrule
-\textbf{2        } &     0.400000 &    0.500000 &    0.500000 &       0.000000 &       0.000000 &     0.500000 &   0.000000 &     0.000000 &  0.000000 &  0.211111 \\
-\textbf{4        } &     0.761410 &    0.600000 &    0.944444 &       0.786636 &       0.729143 &     0.929762 &   0.802857 &     0.941282 &  0.925873 &  0.824601 \\
-\textbf{8        } &     0.724413 &    0.200000 &    1.000000 &       0.775738 &       0.656009 &     0.721905 &   0.813147 &     0.947826 &  0.852976 &  0.743557 \\
-\textbf{16       } &     0.814418 &    0.946913 &    0.927658 &       0.805583 &       0.824701 &     0.911855 &   0.793485 &     0.903535 &  0.911031 &  0.871020 \\
-\textbf{32       } &     0.887372 &    0.865350 &    0.863817 &       0.788800 &       0.841096 &     0.847990 &   0.797868 &     0.933771 &  0.878043 &  0.856012 \\
-\textbf{64       } &     0.896917 &    0.885388 &    0.863708 &       0.761756 &       0.826944 &     0.873175 &   0.829524 &     0.912023 &  0.905253 &  0.861632 \\
-\textbf{128      } &     0.886787 &    0.857769 &    0.841852 &       0.798671 &       0.830715 &     0.858188 &   0.803539 &     0.908465 &  0.864583 &  0.850063 \\
-\textbf{256      } &     0.843112 &    0.858113 &    0.846261 &       0.746803 &       0.762345 &     0.831391 &   0.784388 &     0.933101 &  0.867187 &  0.830300 \\
-\bottomrule
+\resizebox{\textwidth}{!}{
+\begin{tabular}{c|cccccccccc}
+\textbf{Dimension} & \textbf{k\_neighbors} & \textbf{svm\_linear} & \textbf{svm\_radial} & \textbf{decision\_tree} & \textbf{random\_forest} & \textbf{multi\_layer} &\textbf{ada\_boost} &\textbf{gaussian\_nb} &\textbf{ensemble} & \textbf{average} \\ \hline
+\textbf{2} &     0.735390 &    0.200000 &    0.600000 &       0.800320 &       0.689779 &     0.529713 &   0.826404 &     0.913222 &  0.843136 &  0.681996 \\
+\textbf{4} &     0.785431 &    1.000000 &    0.952640 &       0.799541 &       0.776843 &     0.948763 &   0.820187 &     0.954135 &  0.942184 &  0.886636 \\
+\textbf{8} &     0.838084 &    1.000000 &    0.967593 &       0.850238 &       0.868500 &     0.970370 &   0.871131 &     0.956970 &  0.966061 &  0.920994 \\
+\textbf{16} &     0.892353 &    0.983304 &    0.941364 &       0.827188 &       0.854650 &     0.904503 &   0.828368 &     0.896774 &  0.927254 &  0.895084 \\
[…]

[…]
