+++ README.md
@@ -39,7 +39,7 @@ This paper presents our efforts to reproduce and improve the results achieved by
 Clone the repository and the branch:
 
 ```shell
-!git clone --recurse-submodules -j8 https://github.com/bruAristimunha/Re-Deep-Convolution-Neural-Network-and-Autoencoders-Based-Unsupervised-Feature-Learning-of-EEG.git ReScience-submission
+git clone --recurse-submodules -j8 https://github.com/bruAristimunha/Re-Deep-Convolution-Neural-Network-and-Autoencoders-Based-Unsupervised-Feature-Learning-of-EEG.git ReScience-submission
 ```
 
 Install Conda, we recommend the [tutorial](https://docs.conda.io/projects/conda/en/latest/user-guide/install/index.html).
@@ -55,8 +55,10 @@ We recommend using a GPU, if you have it available. But if not, just ignore the 
 ```shell
 conda install -c anaconda tensorflow-gpu 
 pip install -r ReScience-submission/requirements.txt
+cd ReScience-submission/notebook
+jupyter notebook
 ```
 
 ### Execute
 
-We recommend using the [notebook](https://github.com/bruAristimunha/Re-Deep-Convolution-Neural-Network-and-Autoencoders-Based-Unsupervised-Feature-Learning-of-EEG/blob/master/notebook/Jupyter_Paper_Re_Deep_Convolution_Neural_Network_and_Autoencoders_Based_Unsupervised_Feature_Learning_of_EEG_Signals.ipynb) or the colab to reproduce the results.
+Use Jupyter to open the notebook "Jupyter_Paper_Re_Deep_Convolution_Neural_Network_and_Autoencoders_Based_Unsupervised_Feature_Learning_of_EEG_Signals.ipynb" to reproduce the results
