### 2021-07-02 05e6e51
Bump tensorflow-gpu from 2.3.1 to 2.4.2
Bumps [tensorflow-gpu](https://github.com/tensorflow/tensorflow) from 2.3.1 to 2.4.2.
- [Release notes](https://github.com/tensorflow/tensorflow/releases)
- [Changelog](https://github.com/tensorflow/tensorflow/blob/master/RELEASE.md)
- [Commits](https://github.com/tensorflow/tensorflow/compare/v2.3.1...v2.4.2)

---
updated-dependencies:
- dependency-name: tensorflow-gpu
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
