### 2021-07-05 #31 @DiogoEduardo
Hi @gdetor. Sorry about the delay. We completed the changes in the text and adapted the jupyter notebook to work on Linux/Windows. I have tested in both systems using vanilla machine / vanilla conda environment as stated in the readme so I think this type of error is solved to reproduce our results.  The text changes stated before by @bruAristimunha are also completed.
