### 2021-08-25 6c5f718
Bump tensorflow-gpu from 2.3.0 to 2.5.1
Bumps [tensorflow-gpu](https://github.com/tensorflow/tensorflow) from 2.3.0 to 2.5.1.
- [Release notes](https://github.com/tensorflow/tensorflow/releases)
- [Changelog](https://github.com/tensorflow/tensorflow/blob/master/RELEASE.md)
- [Commits](https://github.com/tensorflow/tensorflow/compare/v2.3.0...v2.5.1)

---
updated-dependencies:
- dependency-name: tensorflow-gpu
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
