+++ requirements.txt
@@ -5,7 +5,7 @@ numpy==1.18.2
 pandas==1.0.3
 scikit_learn==0.22.2
 seaborn==0.10.0
-tensorflow-gpu==2.3.0
+tensorflow-gpu==2.5.1
 tqdm==4.38.0
 wget==3.2
 scipy==1.4.1
