### 2021-11-29 #31 @benureau
I am sorry for the delay. Part of it was that I spent a lot of time trying to reproduce the results by re-running training without managing, and thought I was doing something wrong.

In the review below, I detail several problems, big and small, I have with the replication effort right now. However, I have a positive opinion of the work that has been done in this replication under difficult circumstances (ambiguous original article, no possible contact with the original authors, an obvious problem in the original article (copy-pasted error with Figure 3/4)).

## Criterions for replication
Taken from the review guidelines.
### The actual replication of the research. The reviewer must evaluate the authors’ claims about a successful replication, applying the standards of the field.

The authors have done a serious job trying to reproduce the results of the original article. Much of their results fail to reproduce quantitatively.

One problem seems to be that the original article and the authors do not use the same testing dataset for dataset 1. The original article says (p. 25406):
"In addition, because the ratio of positive and negative samples in the testing set is 2:1, the accuracy is 0.667 when all samples were deemed as positive and the accuracy is 0.333 when all samples were deemed as negative."
Yet, in the replication, the ratio seems to be 3:2 (i.e., if I am not mistaken, the whole dataset?), generating scores of 0.6 and 0.4 when all samples are deemed positive or negative respectively (e.g., Table 5 here). Is that something the authors considered? Could the original article only use C, D and E subsets of dataset 1 for testing?

### Reproducibility of the replication. The reviewers must be able to run the proposed implementation on their computer, and obtain the same results as the authors with the limits of the state of the art of the field.

I have tried to rerun the code as much as possible. The replication provides different flags: `download_again`, `train_again` and `eval_again`. The notebook with the code comes precomputed. The results of the precomputed notebook are the same as the article, and we can reproduce the same results by re-running the notebook and the notebook with the flag `eval_again` turned on (`True`). `download_again` is problematic as the Bonn University EEG database seems to have moved or disappeared (any chance to put it in a more stable place on zenodo or figshare?). Downloading the Children’s Hospital of Boston EEG database is slow (big files), and any issue during the download means restarting the download from scratch with the provided code. Not ideal, but works eventually.

I also ran the notebook with `train_again` set to True. I did that on a fully updated Ubuntu 20.04 (as of 2021/11/15), on a NVIDIA RTX 2080ti.
It takes approximately 6 hours. The variance code does not run properly anymore (`KeyError: "['time'] not found in axis"` see below).

I have trouble reproducing the classification results. I sometimes obtain similar results as the ones of the article, but not always. The main problem, though, seems to be that little training is done. This is Figure 9 (notebook numbering):

Am I doing anything wrong? So, right now, the code fails to train for me, it seems. I am attaching the notebook I trained:
[rerun_train.ipynb.zip](https://github.com/ReScience/submissions/files/7620575/rerun_train.ipynb.zip)

### Clarity of the code and the instructions for running it. Uncommented or obfuscated code is as bad as no code at all.
The code is clear enough and the instructions for running it are good.

### Clarity and completeness of the accompanying article, in which the authors should clearly state why they think they have replicated the paper (same figures, same graphics, same behavior, etc.) and explain any obstacles they have encountered during the replication work. The reviewers should also consider if the article is sufficiently self-contained.

I have a problem with this point. The authors do talk quite in detail about the issue with the reproducibility of the variance analysis, and try (in code) several theories they had for the difference. This is well done.

However, considering the classification results, the authors plot and *describe* the quantitative differences between their replication and the original article, but do not provide much if any explanation, or even possible theories for them (e.g., "change in scipy classifiers implementation"). An instance of this is, "Although the results in original paper also have few variations for the classifiers svm_linear, svm_radial and multi_layer we had no variations in these classifers for the autoencoder AE‐CDNN‐MAAE." (btw classifers -> classifiers). This is troubling because in Table 9, `svm_linear` and `svm_radial` fare quite poorly compared to the other classifiers in the replication, with the result of the original paper's Table 9 placing SVM classifier quite high in performance (93% vs 60% in the replication). What is going on there? The features seems to be learned for some classifiers but not others.

It's also not clear if there are qualitative differences between the replication and the original article. Does the replication support the same conclusions of the original article? Some of the text seems to approach this conclusion ("In Figure 9 we observe similar average accuracy between AE‐CDNN‐MAE, AE‐CDNN‐MAAE, PCA and SRP for both datasets. The same is observed in the original article, as well as a similar behavior"), but a more explicit phrasing would be preferable.

## Issues with the replication
- As mentioned before, the apparent 3:2 ratio of the testing dataset for dataset 1, where the original article uses 2:1.

- When `eval_train` is set to True in the notebook, I obtain in section 8.1, first code cell:
```python
KeyError                                  Traceback (most recent call last)
/tmp/ipykernel_3745/1033232197.py in <module>
----> 1 variance_by_file = get_variance_by_file(PATH_CHBMIT)

~/projects/Rescience2021/rerun_train/code/variance.py in get_variance_by_file(path_dataset, range_)
    211                 rank_variance.append(rank_by_file)
    212
--> 213         variance_df = DataFrame(rank_variance).drop("time", 1)
    214
    215         variance = DataFrame(

~/.pyenv/versions/miniconda3-latest/envs/eeg/lib/python3.7/site-packages/pandas/core/frame.py in drop(self, labels, axis, index, columns, level, inplace, errors)
   3995             level=level,
   3996             inplace=inplace,
-> 3997             errors=errors,
   3998         )
   3999

~/.pyenv/versions/miniconda3-latest/envs/eeg/lib/python3.7/site-packages/pandas/core/generic.py in drop(self, labels, axis, index, columns, level, inplace, errors)
   3934         for axis, labels in axes.items():
   3935             if labels is not None:
-> 3936                 obj = obj._drop_axis(labels, axis, level=level, errors=errors)
   3937
   3938         if inplace:

~/.pyenv/versions/miniconda3-latest/envs/eeg/lib/python3.7/site-packages/pandas/core/generic.py in _drop_axis(self, labels, axis, level, errors)
   3968                 new_axis = axis.drop(labels, level=level, errors=errors)
   3969             else:
-> 3970                 new_axis = axis.drop(labels, errors=errors)
   3971             result = self.reindex(**{axis_name: new_axis})
   3972

~/.pyenv/versions/miniconda3-latest/envs/eeg/lib/python3.7/site-packages/pandas/core/indexes/base.py in drop(self, labels, errors)
   5015         if mask.any():
   5016             if errors != "ignore":
-> 5017                 raise KeyError(f"{labels[mask]} not found in axis")
   5018             indexer = indexer[~mask]
   5019         return self.delete(indexer)

KeyError: "['time'] not found in axis"
```

- Failure to train with the provided code.

- The article and the code Figure number do not match. Quite confusing, when we have to deal with the difference with the original article and reproduction article different numbering schemes already. I would suggest switching to manual numbering for everything and number all figures the same as the original article to reduce the cognitive load on the reader. In any case, the notebook and article number should match.

- In Figure 8 (article numbering), the scale of the y axis is orders of magnitude different for AE-CDNN-MAAE between the replication and the original article. Why?

- During setup:
On Ubuntu 20.04 2021/09/21
```
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
tensorflow 2.2.0 requires tensorboard<2.3.0,>=2.2.0, but you have tensorboard 2.6.0 which is incompatible.
tensorflow 2.2.0 requires tensorflow-estimator<2.3.0,>=2.2.0, but you have tensorflow-estimator 2.3.0 which is incompatible.
```

## Minor comments about the text
- "In addition, we have also used/implemented one more loss function, Mean Absolute Average Error ‐ MAAE. The formula presented in the original article by [1] differs from the Mean Absolute Percentage Error ‐ MAPE formula, despite having similar intuitions. Thus, we chose to implement this loss function, and we have not found its use elsewhere, the difference between the loss functions is only in the fact that ..." : confusing text, we don't know what "this loss function" refers too (MAPE or MAAE?). Needs multiple re-read from the reader, and a bit of guesswork to make sure we understood right.

- "Note that we refer to [1]’s AE‐CDNN‐L1 as Loss­MAE and to [1]’s AE‐CDNN‐L2 as Loss­ MAAE" -> while you can point at an easier nomenclature being possibly better, I am not sure it makes it easier for the reader to understand the replication. This sentence also conflates the network and the loss function, which, given all the subsequent mentions to AE-CDNN-MAE and AE-CDNN-MAAE, could be avoided. In the notebook, some mentions of L1 instead of MAE remain.

- "The results obtained can be seen in the Figure below:" -> there is no figure and the text continues (because of Latex float placement). Refer to figures by their number to avoid any ambiguity. Same for tables, e. g. when you say: "in Tables below,".

- "These recordings underwent a pre‐processing in which the signals had a band filter between 0.53 to 40 Hz. There was also the removal of artifacts such as muscle movements or flicker movements." -> The article does not make it clear: you reproduce this pre-processing or was it done for you? How did you remove artifacts?

- "We adopted a Auto‐Encoder that allowed us to construct a smaller representation space." -> It is confusing, in the context of the replication. We is you or the original authors?

- "Table 8. Accuracy in Classification, with loss MAAE, on each fold cross‐validation, for Dataset 2." is not consistent with the captions of Table 6 and 7.

- "The selection followed the methodology used in [9], which analyses the variance of each patient, and after that, chooses the channel of greater variance to represent that individual" -> you make it sound that different individuals are represented by different channels. Is that intended?

- "Bonn" is named "Boon" throughout all the code.

- Checking variance: you explore three scenarios (and I like that a lot). One is named "full database", which is quite confusing to the reader: I was expecting that you would consider the full 23 patients. But only the first 10 are considered in any analysis. Would analysing the 23 patients dataset reproduce the original paper's findings?

## Optional comments about the text:
"Although the efforts in the last years" -> "Despite recent efforts/advances, "?
"are used windows" -> "windows are used"?
"with then" -> "with them"?
"4097 column." -> 	"4097 columns."
