### 2022-01-04 3d3cda5
Items 2-7 and 9.
