### 2022-05-23 #31 @bruAristimunha
Yes, I believe it is the best way to close this issue. I tried several ways to try to fix the reproducibility issues, but unfortunately I couldn't. It seems that choosing jupyter notebook was a bad idea. Unfortunately, I don't have time to finalize all the reviewers' considerations. I apologize profusely for this situation.
